Cool

+++ db/migrate/20180718215119_create_sailboats.rb
@@ -7,7 +7,7 @@ class CreateSailboats < ActiveRecord::Migration[5.2]
       t.string :condition
       t.string :finish
       t.string :title
-      t.integer :price, default: 0
+      t.numeric :price, default: 0
 
       t.timestamps
     end
